Move wrap() function to a separated file in order to wrap opened files.
