+++ aiofiles/__init__.py
@@ -1,7 +1,7 @@
 """Utilities for asyncio-friendly file handling."""
 from .threadpool import open
 from .aiotempfile import NamedAioTemporaryFile, AioTemporaryFile
-
+from .wrap import wrap
 __version__ = '0.3.2'
 
-__all__ = (open, NamedAioTemporaryFile, AioTemporaryFile)
+__all__ = (open, NamedAioTemporaryFile, AioTemporaryFile, wrap)
