+++ aiofiles/threadpool/__init__.py
@@ -1,14 +1,8 @@
 """Handle files using a thread pool executor."""
 import asyncio
-
-from io import (FileIO, TextIOBase, BufferedReader, BufferedWriter,
-                BufferedRandom)
 from functools import partial
-
-from .binary import AsyncBufferedIOBase, AsyncBufferedReader, AsyncFileIO
-from .text import AsyncTextIOWrapper
 from ..base import AiofilesContextManager
-from .._compat import singledispatch
+from ..wrap import wrap
 
 sync_open = open
 
@@ -36,29 +30,3 @@ def _open(file, mode='r', buffering=-1, encoding=None, errors=None, newline=None
     f = yield from loop.run_in_executor(executor, cb)
 
     return wrap(f, loop=loop, executor=executor)
-
-
-@singledispatch
-def wrap(file, *, loop=None, executor=None):
-    raise TypeError('Unsupported io type: {}.'.format(file))
-
-
-@wrap.register(TextIOBase)
-def _(file, *, loop=None, executor=None):
-    return AsyncTextIOWrapper(file, loop=loop, executor=executor)
-
-
-@wrap.register(BufferedWriter)
-def _(file, *, loop=None, executor=None):
-    return AsyncBufferedIOBase(file, loop=loop, executor=executor)
-
-
-@wrap.register(BufferedReader)
-@wrap.register(BufferedRandom)
-def _(file, *, loop=None, executor=None):
-    return AsyncBufferedReader(file, loop=loop, executor=executor)
-
-
-@wrap.register(FileIO)
-def _(file, *, loop=None, executor=None):
-    return AsyncFileIO(file, loop, executor)
